+++ src/main/java/nextstep/subway/path/domain/fare/DistanceCalculator.java
@@ -1,4 +1,4 @@
-package nextstep.subway.fare.domain;
+package nextstep.subway.path.domain.fare;
 
 import java.util.Arrays;
