+++ src/test/java/nextstep/subway/path/domain/fare/FareBuilderTest.java
@@ -1,4 +1,4 @@
-package nextstep.subway.fare.domain;
+package nextstep.subway.path.domain.fare;
 
 import static org.assertj.core.api.Assertions.*;
